refactor(server): collapse the base/impl deployment-manager abstraction
`BaseDeploymentManager` was an abstract class with ~7 abstract methods
"to work with any deployment backend", but it had exactly one concrete
implementor (`EmbeddedDeploymentManager`) and `Orchestrator` hardcoded
`new EmbeddedDeploymentManager(...)` with no backend selection. The
anticipated K8s/cloud backend was never built.

Fold the abstract base into a single concrete `DeploymentManager`
(renamed from `EmbeddedDeploymentManager`, kept at
`orchestration/deployment-manager.ts`) and delete the dead abstraction:

- Merge `base-deployment-manager.ts` into the (renamed) impl file; drop
  the `abstract` class + abstract-method indirection. The previously
  abstract methods (`spawnDeployment`, `listDeployments`,
  `scaleDeployment`, `deleteDeployment`, `updateDeploymentActivity`,
  `validateWorkerImage`, `getDispatcherHost`) are now their concrete
  embedded implementations. The base's `getDispatcherUrl` (port 8080)
  was always overridden by the embedded `/lobu`-mounted version, so only
  the embedded one survives — behavior is identical.
- Delete `base-deployment-manager.ts` and the now-empty `impl/` dir.
- Update `index.ts` + all call sites/imports (config, message-consumer,
  deployment-utils, src/index) and tests to the single class/module.

No shim, no @deprecated alias. All exported helpers/types are preserved
verbatim. Behavior is byte-identical; this only removes the indirection.

Did NOT extract WorkerEnvironmentBuilder / ConfigurationSyncer — that
reshapes the dependency wiring and risks behavior drift on the worker
orchestration path; out of scope for a pure collapse.
