+++ packages/agent-worker/src/gateway/sse-client.ts
@@ -509,7 +509,7 @@ export class GatewayClient {
 
     // No per-user filtering here: deployment names intentionally hash only
     // `platform:channelId:conversationId` (see `generateDeploymentName` in
-    // base-deployment-manager.ts) so a channel/thread has ONE shared worker
+    // deployment-manager.ts) so a channel/thread has ONE shared worker
     // across all posting users. DMs are single-participant, so a check would
     // be dead there too. The WORKER_TOKEN-scoped-to-spawning-user tradeoff
     // for shared channel workers is acknowledged and deferred to per-message
